Refactor tests to group by error number.
Closes #26.  Separate modules for each linter error number, including all corresponding test signatures.
